(#70) Remove usage of Language Server
Simply put, this isn't ready for usage, and lots of decisions need to be
made about how it will be used going forward.  Rather than prevent other
releases of this extension while we are waiting on the Language Server
support, let's move the current code into a branch and continue over
there.