(#70) Add current Language Server code
This code has been taken from the develop branch, so that we can
continue the development of the language server portion of the extension
without preventing other ongoing work from going out the door.

We will get to this at some point, but for now, this should be regarded
as a nice have.